+++ src/hooks/usePromptBuilder.ts
@@ -153,9 +153,6 @@ export function usePromptBuilder() {
 	}, [modifiedPromptText, worldInfo]);
 
 	const additionalContextPrompt = useMemo(() => {
-		// add world info to memory for easier assembly
-		memoryTokens["worldInfo"] = assembledWorldInfo;
-
 		const order: (keyof AuthorNoteData)[] = ["prefix","text","suffix"]
 		const assembledAuthorNote = authorNoteTokens.text && authorNoteTokens.text !== ""
 			? order.map(key => authorNoteTokens[key]).join("").replace(/\\n/g,'\n')